Further work in implementation by using atomic level compilation and diffing
