+++ pkg/quarks/extractors/extractors.go
@@ -2,3 +2,6 @@ package extractors
 
 // Extractor is a extractor structure parsed from a yaml file
 type Extractor struct{}
+
+// CompiledExtractor is the compiled extractor parsed from yaml file.
+type CompiledExtractor struct{}
