+++ pkg/quarks/matchers/matchers.go
@@ -2,3 +2,6 @@ package matchers
 
 // Matcher is a matcher structure parsed from a yaml file
 type Matcher struct{}
+
+// CompiledMatcher is the compiled matcher parsed from yaml file.
+type CompiledMatcher struct{}
