+++ pkg/quarks/templates/templates.go
@@ -1,9 +1,8 @@
 package templates
 
 import (
-	"net/http"
-
 	"github.com/projectdiscovery/nuclei/v2/pkg/quarks/requests/dns"
+	"github.com/projectdiscovery/nuclei/v2/pkg/quarks/requests/http"
 )
 
 // Template is a request workflow parsed from a yaml file
@@ -15,4 +14,10 @@ type Template struct {
 
 	// DNS contains the DNS requests to make to the targets.
 	DNS []dns.Request `yaml:"dns"`
+}
+
+// CompiledTemplate is the compiled template workflow parsed from yaml file.
+type CompiledTemplate struct {
+	DNS  []dns.CompiledRequest
+	HTTP []http.CompiledRequest
 }
