+++ pkg/requests/dns-request.go
@@ -1,8 +1,6 @@
 package requests
 
 import (
-	"strings"
-
 	"github.com/miekg/dns"
 	"github.com/projectdiscovery/nuclei/v2/pkg/extractors"
 	"github.com/projectdiscovery/nuclei/v2/pkg/matchers"
@@ -67,55 +65,4 @@ func (r *DNSRequest) MakeDNSRequest(domain string) (*dns.Msg, error) {
 	req.Question = append(req.Question, q)
 
 	return req, nil
-}
-
-func toQType(ttype string) (rtype uint16) {
-	ttype = strings.TrimSpace(strings.ToUpper(ttype))
-
-	switch ttype {
-	case "A":
-		rtype = dns.TypeA
-	case "NS":
-		rtype = dns.TypeNS
-	case "CNAME":
-		rtype = dns.TypeCNAME
-	case "SOA":
-		rtype = dns.TypeSOA
-	case "PTR":
-		rtype = dns.TypePTR
-	case "MX":
-		rtype = dns.TypeMX
-	case "TXT":
-		rtype = dns.TypeTXT
-	case "AAAA":
-		rtype = dns.TypeAAAA
-	default:
-		rtype = dns.TypeA
-	}
-
-	return
-}
-
-func toQClass(tclass string) (rclass uint16) {
-	tclass = strings.TrimSpace(strings.ToUpper(tclass))
-
-	switch tclass {
-	case "INET":
-		rclass = dns.ClassINET
-	case "CSNET":
-		rclass = dns.ClassCSNET
-	case "CHAOS":
-		rclass = dns.ClassCHAOS
-	case "HESIOD":
-		rclass = dns.ClassHESIOD
-	case "NONE":
-		rclass = dns.ClassNONE
-	case "ANY":
-		rclass = dns.ClassANY
-	default:
-		// Use INET by default.
-		rclass = dns.ClassINET
-	}
-
-	return
 }
